refactor: move and rename connection logs in runtime logs details
