+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/components/api-runtime-logs-connection-log-details/api-runtime-logs-connection-log-details.component.html
