+++ gravitee-apim-console-webui/src/management/api/runtime-logs-v4/runtime-logs-details/components/api-runtime-logs-connection-log-details/api-runtime-logs-connection-log-details.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '../../../../../../../../node_modules/@angular/material/index' as mat;
-@use '../../../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
+@use '../../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 $typography: map.get(gio.$mat-theme, typography);
